Refactoring Completed
