+++ package.json
@@ -19,12 +19,21 @@
   },
   "homepage": "https://github.com/woowahan-techcamp/3team_OSGD.fling#readme",
   "devDependencies": {
-    "babel-core": "^6.25.0",
+    "babel-core": "^6.26.0",
     "babel-loader": "^7.1.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-es2017": "^6.24.1",
     "css-loader": "^0.28.4",
     "style-loader": "^0.18.2",
     "webpack": "^3.5.3",
     "webpack-dev-server": "^2.7.1"
+  },
+  "dependencies": {
+    "babel-runtime": "^6.26.0",
+    "handlebars": "^4.0.10",
+    "handlebars-loader": "^1.5.0"
   }
 }
