+++ src/css/refrige_popup.css
@@ -142,6 +142,7 @@
     height: 50px;
     padding: 20px 60px 20px 150px;
     border-bottom: 1px solid #25d4ce;
+    transition: opacity 2s;
 }
 
 .material_name {
