+++ src/js/Utils.js
@@ -1,4 +1,5 @@
-class Utils {
+window.Fling = window.Fling || {};
+window.Fling.Utils = class Utils {
     static getParameterByName(name, url = window.location.href) {
         name = name.replace(/[\[\]]/g, "\\$&");
         let regex = new RegExp(`[?&]${name}(=([^&$]*)|&|#|$)`),
@@ -26,5 +27,21 @@ class Utils {
 
     static numberWithComma(x) {
         return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+    }
+
+    static PopupCenter(url, title, w, h) {
+        var dualScreenLeft = window.screenLeft != undefined ? window.screenLeft : screen.left;
+        var dualScreenTop = window.screenTop != undefined ? window.screenTop : screen.top;
+    
+        var width = window.innerWidth ? window.innerWidth : document.documentElement.clientWidth ? document.documentElement.clientWidth : screen.width;
+        var height = window.innerHeight ? window.innerHeight : document.documentElement.clientHeight ? document.documentElement.clientHeight : screen.height;
+    
+        var left = ((width / 2) - (w / 2)) + dualScreenLeft;
+        var top = ((height / 2) - (h / 2)) + dualScreenTop;
+        var newWindow = window.open(url, title, 'scrollbars=yes, width=' + w + ', height=' + h + ', top=' + top + ', left=' + left);
+    
+        if (window.focus) {
+            newWindow.focus();
+        }
     }
 }
