+++ src/recipe_page.html
@@ -5,11 +5,7 @@
         <title>플링</title>
         <link rel="stylesheet" href="./css/common.css">
         <link rel="stylesheet" href="./css/recipe_page.css">
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.10/handlebars.js"></script>
-        <script src="js/Fling.js"></script>
-        <script src="./js/XHR.js"></script>
-        <script src="./js/Utils.js"></script>
-        <script src="./recipe_page.js"></script>
+        <script src="dist/bundle.js"></script>
         <script>
         document.addEventListener("DOMContentLoaded", Fling.RecipePage.EventHandler);
         </script>
