+++ src/recipe_popup.html
@@ -4,11 +4,10 @@
         <title>플링 :: 레시피 상세정보</title>
         <link rel="stylesheet" href="./css/common.css">
         <link rel="stylesheet" href="./css/recipe_popup.css">
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.10/handlebars.js"></script>
-        <script src="js/Fling.js"></script>
-        <script src="./js/XHR.js"></script>
-        <script src="./js/Utils.js"></script>
-        <script src="./recipe_popup.js"></script>
+        <script src="dist/bundle.js"></script>
+        <script>
+        document.addEventListener('DOMContentLoaded', Fling.RecipePopup.EventHandler);
+        </script>
         <script type="Template" id="product_list_template">
             {{#each this.productDetail}}
                 {{#if isChecked}}
